pyproject.toml
[tool.poetry]
name = "crossbar_llm"
version = "0.1.0"
description = ""
authors = ["Melih DARCAN <57872471+MelihDarcanxyz@users.noreply.github.com>", "Bünyamin Şen <106750021+bnymnsen@users.noreply.github.com>"]
license = "License :: OSI Approved :: GNU General Public License v3 (GPLv3)"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"

langchain = "^0.3.0"
langchain-anthropic = "^0.2.3"
langchain-community = "^0.3.0"
langchain-google-genai = "^2.0.1"
langchain-groq = "^0.2.0"
neo4j = "^5.20.0"
numpy = "^1.26.4"
openai = "^1.30.2"
pandas = "^2.2.2"
pydantic = "^2.7.1"
python-dotenv = "^1.0.1"
langchain-nvidia-ai-endpoints = "^0.3.0"
wrapt-timeout-decorator = "^1.5.1"
langchain-openai = "^0.3.0"
fastapi = "^0.115.5"
fastapi-csrf-protect = "0.3.2"
python-multipart = "^0.0.17"
sse-starlette = "^2.2.1"
uvicorn = "^0.35.0"

[tool.poetry.group.dev.dependencies]

[tool.poetry-auto-export]
output = "requirements.txt"
without_hashes = true
without = ["dev"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
